[agentless-scanner] Convert agentless-scanner command to the subcommands hierarchy
